refactor: move catalog layout components in 'layout/catalog' folder
